requirements.txt
jinja2==3.1.2
python-dotenv==0.20.0
gunicorn==20.1.0
Flask==2.1.3
spotipy==2.20.0
pymongo==4.2.0
dnspython==2.2.1